Use AtomicLong

+++ library/runtime-ctrl/src/jvmMain/kotlin/io/matthewnelson/kmp/tor/runtime/ctrl/internal/-JvmPlatform.kt
@@ -28,7 +28,6 @@ import io.matthewnelson.kmp.tor.runtime.ctrl.TorCtrl
 import kotlinx.coroutines.CloseableCoroutineDispatcher
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.asCoroutineDispatcher
-import kotlinx.coroutines.newFixedThreadPoolContext
 import java.io.InputStream
 import java.io.OutputStream
 import java.lang.reflect.Constructor
@@ -43,12 +42,12 @@ import java.nio.channels.ReadableByteChannel
 import java.nio.channels.SocketChannel
 import java.nio.channels.WritableByteChannel
 import java.util.concurrent.Executors
-import java.util.concurrent.atomic.AtomicInteger
+import java.util.concurrent.atomic.AtomicLong
 import kotlin.concurrent.Volatile
 
 @OptIn(ExperimentalCoroutinesApi::class)
 internal actual fun TorCtrl.Factory.newTorCtrlDispatcher(): CloseableCoroutineDispatcher {
-    val threadNo = AtomicInteger()
+    val threadNo = AtomicLong()
     val executor = Executors.newFixedThreadPool(2) { runnable ->
         val t = Thread(runnable, "TorCtrl-${threadNo.incrementAndGet()}")
         t.isDaemon = true
